ci: give up on being able to generate bindings on macOS
This isn't really of any use anyways and I suspect this is brittle enough that even if I get things working on my version of macOS with my configuration it's still not very likely to work on other machines.
Bindings can be generated on ubuntu and debian in CI and on my local Linux installation. I don't expect that anyone will actually want to generate the bindings themselves anyways, so I'm going to call this good enough for now and move on.