bindings: fix the windows bad_alloc bug by actually freeing memory..
I'm not convinced this is actually the problem; the fact that it
repeatedly tries to allocate ~500MB (540028976 bytes) is highly suspect
and maybe indicates that somehow the OS source is getting corrupted or
something is breaking (printing out the OS source repeatedly did not
reveal anything).

But, this'll do for now.